[REFACTOR] Refactor submodules into blueprints
