+++ build-tools/github/validate-python
@@ -10,4 +10,4 @@ echo "------> Linting Python code"
 # Let's just check for definite errors.
 pylint -j 0 \
     --errors-only \
-    *.py tests
+    *.py website highlighting tests
