+++ requirements.txt
@@ -12,7 +12,7 @@ ruamel.yaml==0.17.4
 docopt==0.6.2
 pydantic==1.8.2
 lazy==1.4
-pylint==2.8.2
+pylint==2.15.3
 awscli>=1.19.88
 PySumTypes==0.0.1
 beautifulsoup4==4.9.3
