+++ website/admin.py
@@ -1,26 +1,30 @@
 from flask_babel import gettext
+from flask import request, g
 import hedyweb
 from website import statistics
-from website.auth import requires_login, current_user, is_admin, pick, requires_admin
+from website.auth import create_verify_link, current_user, is_admin, is_teacher, make_salt, pick, requires_admin, send_email_template
+from .database import Database
 import utils
-from flask import request
 from flask_helpers import render_template
+from .website_module import WebsiteModule, route
 
 
-def routes(app, database):
-    global DATABASE
-    DATABASE = database
+class AdminModule(WebsiteModule):
+    def __init__(self, db: Database):
+        super().__init__('admin', __name__, url_prefix='admin')
 
-    @app.route('/admin', methods=['GET'])
-    def get_admin_page():
+        self.db = db
+
+    @route('/', methods=['GET'])
+    def get_admin_page(self):
         # Todo TB: Why do we check for the testing_request here? (09-22)
         if not utils.is_testing_request(request) and not is_admin(current_user()):
             return utils.error_page(error=403, ui_message=gettext('unauthorized'))
         return render_template('admin/admin.html', page_title=gettext('title_admin'))
 
-    @app.route('/admin/users', methods=['GET'])
+    @route('/users', methods=['GET'])
     @requires_admin
-    def get_admin_users_page(user):
+    def get_admin_users_page(self, user):
         category = request.args.get('filter', default=None, type=str)
         category = None if category == "null" else category
 
@@ -38,7 +42,7 @@ def routes(app, database):
 
         pagination_token = request.args.get('page', default=None, type=str)
 
-        users = DATABASE.all_users(pagination_token)
+        users = g.db.all_users(pagination_token)
 
         userdata = []
         fields = [
@@ -83,13 +87,13 @@ def routes(app, database):
             userdata.append(data)
 
         return render_template('admin/admin-users.html', users=userdata, page_title=gettext('title_admin'),
-                               filter=category, start_date=start_date, end_date=end_date, text_filter=substring,
-                               language_filter=language, keyword_language_filter=keyword_language,
-                               next_page_token=users.next_page_token)
+                                filter=category, start_date=start_date, end_date=end_date, text_filter=substring,
+                                language_filter=language, keyword_language_filter=keyword_language,
+                                next_page_token=users.next_page_token)
 
-    @app.route('/admin/classes', methods=['GET'])
+    @route('/classes', methods=['GET'])
     @requires_admin
-    def get_admin_classes_page(user):
+    def get_admin_classes_page(self, user):
         classes = [{
             "name": Class.get('name'),
             "teacher": Class.get('teacher'),
@@ -97,16 +101,16 @@ def routes(app, database):
             "students": len(Class.get('students')) if 'students' in Class else 0,
             "stats": statistics.get_general_class_stats(Class.get('students', [])),
             "id": Class.get('id')
-        } for Class in DATABASE.all_classes()]
+        } for Class in g.db.all_classes()]
 
         classes = sorted(classes, key=lambda d: d.get('stats').get('week').get('runs'), reverse=True)
 
         return render_template('admin/admin-classes.html', classes=classes, page_title=gettext('title_admin'))
 
-    @app.route('/admin/adventures', methods=['GET'])
+    @route('/adventures', methods=['GET'])
     @requires_admin
-    def get_admin_adventures_page(user):
-        all_adventures = sorted(DATABASE.all_adventures(), key=lambda d: d.get('date', 0), reverse=True)
+    def get_admin_adventures_page(self, user):
+        all_adventures = sorted(g.db.all_adventures(), key=lambda d: d.get('date', 0), reverse=True)
         adventures = [{
             "id": adventure.get('id'),
             "creator": adventure.get('creator'),
@@ -119,19 +123,19 @@ def routes(app, database):
 
         return render_template('admin/admin-adventures.html', adventures=adventures, page_title=gettext('title_admin'))
 
-    @app.route('/admin/stats', methods=['GET'])
+    @route('/stats', methods=['GET'])
     @requires_admin
-    def get_admin_stats_page(user):
+    def get_admin_stats_page(self, user):
         return render_template('admin/admin-stats.html', page_title=gettext('title_admin'))
 
-    @app.route('/admin/logs', methods=['GET'])
+    @route('/logs', methods=['GET'])
     @requires_admin
-    def get_admin_logs_page(user):
+    def get_admin_logs_page(self, user):
         return render_template('admin/admin-logs.html', page_title=gettext('title_admin'))
 
-    @app.route('/admin/achievements', methods=['GET'])
+    @route('/achievements', methods=['GET'])
     @requires_admin
-    def get_admin_achievements_page(user):
+    def get_admin_achievements_page(self, user):
         stats = {}
         achievements = hedyweb.AchievementTranslations().get_translations("en").get("achievements")
         for achievement in achievements.keys():
@@ -140,12 +144,123 @@ def routes(app, database):
             stats[achievement]["description"] = achievements.get(achievement).get("text")
             stats[achievement]["count"] = 0
 
-        user_achievements = DATABASE.get_all_achievements()
+        user_achievements = g.db.get_all_achievements()
         total = len(user_achievements)
         for user in user_achievements:
             for achieved in user.get("achieved", []):
                 stats[achieved]["count"] += 1
 
         return render_template('admin/admin-achievements.html', stats=stats,
-                               total=total, page_title=gettext('title_admin'))
+                                total=total, page_title=gettext('title_admin'))
+
+
+    @route('/markAsTeacher', methods=['POST'])
+    def mark_as_teacher(self):
+        user = current_user()
+        if not is_admin(user) and not utils.is_testing_request(request):
+            return utils.error_page(error=403, ui_message=gettext('unauthorized'))
+
+        body = request.json
+
+        # Validations
+        if not isinstance(body, dict):
+            return gettext('ajax_error'), 400
+        if not isinstance(body.get('username'), str):
+            return gettext('username_invalid'), 400
+        if not isinstance(body.get('is_teacher'), bool):
+            return gettext('teacher_invalid'), 400
+
+        user = g.db.user_by_username(body['username'].strip().lower())
+
+        if not user:
+            return gettext('username_invalid'), 400
+
+        is_teacher_value = 1 if body['is_teacher'] else 0
+        update_is_teacher(self.db, user, is_teacher_value)
+
+        # Todo TB feb 2022 -> Return the success message here instead of fixing in the front-end
+        return '', 200
+
+    @route('/changeUserEmail', methods=['POST'])
+    @requires_admin
+    def change_user_email(self, user):
+        body = request.json
+
+        # Validations
+        if not isinstance(body, dict):
+            return gettext('ajax_error'), 400
+        if not isinstance(body.get('username'), str):
+            return gettext('username_invalid'), 400
+        if not isinstance(body.get('email'), str) or not utils.valid_email(body['email']):
+            return gettext('email_invalid'), 400
+
+        user = g.db.user_by_username(body['username'].strip().lower())
+
+        if not user:
+            return gettext('email_invalid'), 400
+
+        token = make_salt()
+        hashed_token = hash(token, make_salt())
+
+        # We assume that this email is not in use by any other users. In other words, we trust the admin to enter a valid, not yet used email address.
+        g.db.update_user(user['username'], {'email': body['email'], 'verification_pending': hashed_token})
+
+        # If this is an e2e test, we return the email verification token directly instead of emailing it.
+        if utils.is_testing_request(request):
+            resp = {'username': user['username'], 'token': hashed_token}
+        else:
+            try:
+                send_email_template(template='welcome_verify', email=body['email'],
+                                    link=create_verify_link(user['username'], hashed_token),
+                                    username=user['username'])
+            except:
+                return gettext('mail_error_change_processed'), 400
+
+        return {}, 200
+
+    @route('/getUserTags', methods=['POST'])
+    @requires_admin
+    def get_user_tags(self, user):
+        body = request.json
+        user = g.db.get_public_profile_settings(body['username'].strip().lower())
+        if not user:
+            return "User doesn't have a public profile", 400
+        return {'tags': user.get('tags', [])}, 200
+
+    @route('/updateUserTags', methods=['POST'])
+    @requires_admin
+    def update_user_tags(self, user):
+        body = request.json
+        user = g.db.get_public_profile_settings(body['username'].strip().lower())
+        if not user:
+            return "User doesn't have a public profile", 400
+
+        tags = []
+        if "admin" in user.get('tags', []):
+            tags.append("admin")
+        if "teacher" in user.get('tags', []):
+            tags.append("teacher")
+        if body.get("certified"):
+            tags.append("certified_teacher")
+        if body.get("distinguished"):
+            tags.append("distinguished_user")
+        if body.get("contributor"):
+            tags.append("contributor")
+
+        user['tags'] = tags
+        g.db.update_public_profile(user['username'], user)
+        return {}, 200
+
+
+def update_is_teacher(db: Database, user, is_teacher_value=1):
+    user_is_teacher = is_teacher(user)
+    user_becomes_teacher = is_teacher_value and not user_is_teacher
+
+    db.update_user(user['username'], {'is_teacher': is_teacher_value, 'teacher_request': None})
+
+    if user_becomes_teacher and not utils.is_testing_request(request):
+        try:
+            send_email_template(template='welcome_teacher', email=user['email'], username=user['username'])
+        except:
+            print(f"An error occurred when sending a welcome teacher mail to {user['email']}, changes still processed")
