+++ website/auth.py
@@ -1,35 +1,31 @@
-import ast
-import collections
+import json
 import os
-from flask_babel import gettext
-import utils
-from hedy_content import COUNTRIES, ALL_LANGUAGES, ALL_KEYWORD_LANGUAGES
-from website.yaml_file import YamlFile
 import bcrypt
 import re
 import urllib
-from flask import request, session, make_response, jsonify, redirect
-from utils import timems, times, extract_bcrypt_rounds, is_testing_request, is_debug_mode, valid_email, is_heroku
-import datetime
 from functools import wraps
-from config import config
+
 import boto3
 from botocore.exceptions import ClientError as email_error, NoCredentialsError
-import json
+from flask import g, request, session
+from flask_babel import gettext
 import requests
-from website import querylog, database
-import hashlib
+
+import utils
+from utils import timems, times, is_debug_mode
+from config import config
+from website import querylog
 
 TOKEN_COOKIE_NAME = config['session']['cookie_name']
+
 # The session_length in the session is set to 60 * 24 * 14 (in minutes) config.py#13
 # The reset_length in the session is set to 60 * 4 (in minutes) config.py#14
 # We multiply this by 60 to set the session_length to 14 days and reset_length to 4 hours
-session_length = config['session']['session_length'] * 60
-reset_length = config['session']['reset_length'] * 60
+SESSION_LENGTH = config['session']['session_length'] * 60
+RESET_LENGTH = config['session']['reset_length'] * 60
+
 
 env = os.getenv('HEROKU_APP_NAME')
-
-DATABASE: database.Database = None
 
 MAILCHIMP_API_URL = None
 if os.getenv('MAILCHIMP_API_KEY') and os.getenv('MAILCHIMP_AUDIENCE_ID'):
@@ -101,7 +97,7 @@ def current_user():
     if ttl == None or now >= ttl:
         username = user['username']
         if username:
-            db_user = DATABASE.user_by_username(username)
+            db_user = g.db.user_by_username(username)
             if not db_user:
                 raise RuntimeError(f'Cannot find current user in db anymore: {username}')
             remember_current_user(db_user)
@@ -142,19 +138,6 @@ def is_teacher(user):
     return bool(user.get('is_teacher', False))
 
 
-def update_is_teacher(user, is_teacher_value=1):
-    user_is_teacher = is_teacher(user)
-    user_becomes_teacher = is_teacher_value and not user_is_teacher
-
-    DATABASE.update_user(user['username'], {'is_teacher': is_teacher_value, 'teacher_request': None})
-
-    if user_becomes_teacher and not is_testing_request(request):
-        try:
-            send_email_template(template='welcome_teacher', email=user['email'], username=user['username'])
-        except:
-            print(f"An error occurred when sending a welcome teacher mail to {user['email']}, changes still processed")
-
-
 # Thanks to https://stackoverflow.com/a/34499643
 def requires_login(f):
     @wraps(f)
@@ -184,25 +167,15 @@ def login_user_from_token_cookie():
     if not request.cookies.get(TOKEN_COOKIE_NAME):
         return
 
-    token = DATABASE.get_token(request.cookies.get(TOKEN_COOKIE_NAME))
+    token = g.db.get_token(request.cookies.get(TOKEN_COOKIE_NAME))
     if not token:
         return
 
     # We update the login record with the current time -> this way the last login is closer to correct
-    DATABASE.record_login(token['username'])
-    user = DATABASE.user_by_username(token['username'])
+    g.db.record_login(token['username'])
+    user = g.db.user_by_username(token['username'])
     if user:
         remember_current_user(user)
-
-
-def prepare_user_db(username, password):
-    hashed = hash(password, make_salt())
-
-    token = make_salt()
-    hashed_token = hash(token, make_salt())
-    username = username.strip().lower()
-
-    return username, hashed, hashed_token
 
 
 def validate_student_signup_data(account):
@@ -235,640 +208,6 @@ def validate_signup_data(account):
     return None
 
 
-def store_new_student_account(account, teacher_username):
-    username, hashed, hashed_token = prepare_user_db(account['username'], account['password'])
-    user = {
-        'username': username,
-        'password': hashed,
-        'language': account['language'],
-        'keyword_language': account['keyword_language'],
-        'created': timems(),
-        'teacher': teacher_username,
-        'verification_pending': hashed_token,
-        'last_login': timems()
-    }
-    DATABASE.store_user(user)
-    return user
-
-
-def store_new_account(account, email):
-    username, hashed, hashed_token = prepare_user_db(account['username'], account['password'])
-    user = {
-        'username': username,
-        'password': hashed,
-        'email': email,
-        'language': account['language'],
-        'keyword_language': account['keyword_language'],
-        'created': timems(),
-        'teacher_request': True if account.get('is_teacher') else None,
-        'third_party': True if account.get('agree_third_party') else None,
-        'verification_pending': hashed_token,
-        'last_login': timems()
-    }
-
-    for field in ['country', 'birth_year', 'gender', 'language', 'prog_experience', 'experience_languages']:
-        if field in account:
-            if field == 'experience_languages' and len(account[field]) == 0:
-                continue
-            user[field] = account[field]
-
-    DATABASE.store_user(user)
-
-    # If this is an e2e test, we return the email verification token directly instead of emailing it.
-    if is_testing_request(request):
-        resp = make_response({'username': username, 'token': hashed_token})
-    # Otherwise, we send an email with a verification link and we return an empty body
-    else:
-        try:
-            send_email_template(template='welcome_verify', email=email,
-                            link=create_verify_link(username, hashed_token), username=user['username'])
-        except:
-            return user, make_response({gettext('mail_error_change_processed')}, 400)
-        resp = make_response({})
-    return user, resp
-
-
-def create_recover_link(username, token):
-    email = email_base_url() + '/reset?username='
-    email += urllib.parse.quote_plus(username) + '&token=' + urllib.parse.quote_plus(token)
-    return email
-
-
-def create_verify_link(username, token):
-    email = email_base_url() + '/auth/verify?username='
-    email += urllib.parse.quote_plus(username) + '&token=' + urllib.parse.quote_plus(token)
-    return email
-
-# Note: translations are used only for texts that will be seen by a GUI user.
-def routes(app, database):
-    global DATABASE
-    DATABASE = database
-
-    @app.route('/auth/login', methods=['POST'])
-    def login():
-        body = request.json
-        # Validations
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-        if not isinstance(body.get('password'), str):
-            return gettext('password_invalid'), 400
-
-        # If username has an @-sign, then it's an email
-        if '@' in body['username']:
-            user = DATABASE.user_by_email(body['username'])
-        else:
-            user = DATABASE.user_by_username(body['username'])
-
-        if not user or not check_password(body['password'], user['password']):
-            return gettext('invalid_username_password') + " " + gettext('no_account'), 403
-
-        # If the number of bcrypt rounds has changed, create a new hash.
-        new_hash = None
-        if config['bcrypt_rounds'] != extract_bcrypt_rounds(user['password']):
-            new_hash = hash(body['password'], make_salt())
-
-        cookie = make_salt()
-        DATABASE.store_token({'id': cookie, 'username': user['username'], 'ttl': times() + session_length})
-        if new_hash:
-            DATABASE.record_login(user['username'], new_hash)
-        else:
-            DATABASE.record_login(user['username'])
-
-        # Check if the user has a public profile, if so -> retrieve the profile image
-        public_profile = DATABASE.get_public_profile_settings(user['username'])
-        if public_profile:
-            session['profile_image'] = public_profile.get('image', 1)
-
-        # Make an empty response to make sure we have one
-        resp = make_response()
-
-        if is_admin(user):
-            resp = make_response({'admin': True})
-        elif user.get('is_teacher'):
-            resp = make_response({'teacher': True})
-
-        # If the user is a student (and has a related teacher) and the verification is still pending -> first login
-        if user.get('teacher') and user.get('verification_pending'):
-            DATABASE.update_user(user['username'], {'verification_pending': None})
-            resp = make_response({'first_time': True})
-
-        # We set the cookie to expire in a year, just so that the browser won't invalidate it if the same cookie gets renewed by constant use.
-        # The server will decide whether the cookie expires.
-        resp.set_cookie(TOKEN_COOKIE_NAME, value=cookie, httponly=True, secure=is_heroku(), samesite='Lax', path='/',
-                        max_age=365 * 24 * 60 * 60)
-
-        # Remember the current user on the session. This is "new style" logins, which should ultimately
-        # replace "old style" logins (with the cookie above), as it requires fewer database calls.
-        remember_current_user(user)
-        return resp
-
-    @app.route('/auth/signup', methods=['POST'])
-    def signup():
-        body = request.json
-        # Validations, mandatory fields
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-
-        # Validate the essential data using a function -> also used for multiple account creation
-        validation = validate_signup_data(body)
-        if validation:
-            return validation, 400
-
-        # Validate fields only relevant when creating a single user account
-        if not isinstance(body.get('password_repeat'), str) or body['password'] != body['password_repeat']:
-            return gettext('repeat_match_password'), 400
-        if not isinstance(body.get('language'), str) or body.get('language') not in ALL_LANGUAGES.keys():
-            return gettext('language_invalid'), 400
-        if not isinstance(body.get('agree_terms'), str) or not body.get('agree_terms'):
-            return gettext('agree_invalid'), 400
-        if not isinstance(body.get('keyword_language'), str) or body.get('keyword_language') not in ['en', body.get('language')]:
-            return gettext('keyword_language_invalid'), 400
-
-        # Validations, optional fields
-        if 'birth_year' in body:
-            year = datetime.datetime.now().year
-            try:
-                body['birth_year'] = int(body.get('birth_year'))
-            except ValueError:
-                return gettext('year_invalid').format(**{'current_year': str(year)}), 400
-            if not isinstance(body.get('birth_year'), int) or body['birth_year'] <= 1900 or body['birth_year'] > year:
-                return gettext('year_invalid').format(**{'current_year': str(year)}), 400
-        if 'gender' in body:
-            if body['gender'] != 'm' and body['gender'] != 'f' and body['gender'] != 'o':
-                return gettext('gender_invalid'), 400
-        if 'country' in body:
-            if not body['country'] in COUNTRIES:
-                return gettext('country_invalid'), 400
-        if 'prog_experience' in body and body['prog_experience'] not in ['yes', 'no']:
-            return gettext('experience_invalid'), 400
-        if 'experience_languages' in body:
-            if not isinstance(body['experience_languages'], list):
-                return gettext('experience_invalid'), 400
-            for language in body['experience_languages']:
-                if language not in['scratch', 'other_block', 'python', 'other_text']:
-                    return gettext('programming_invalid'), 400
-
-        if DATABASE.user_by_username(body['username'].strip().lower()):
-            return gettext('exists_username'), 403
-        if DATABASE.user_by_email(body['email'].strip().lower()):
-            return gettext('exists_email'), 403
-
-        # We receive the pre-processed user and response package from the function
-        user, resp = store_new_account(body, body['email'].strip().lower())
-
-        if not is_testing_request(request) and 'subscribe' in body and body['subscribe'] is True:
-            # If we have a Mailchimp API key, we use it to add the subscriber through the API
-            if MAILCHIMP_API_URL:
-                mailchimp_subscribe_user(user['email'])
-            # Otherwise, we send an email to notify about the subscription to the main email address
-            else:
-                send_email(config['email']['sender'], 'Subscription to Hedy newsletter on signup', user['email'],
-                           '<p>' + user['email'] + '</p>')
-
-        # We automatically login the user
-        cookie = make_salt()
-        DATABASE.store_token({'id': cookie, 'username': user['username'], 'ttl': times() + session_length})
-        # We set the cookie to expire in a year, just so that the browser won't invalidate it if the same cookie gets renewed by constant use.
-        # The server will decide whether the cookie expires.
-        resp.set_cookie(TOKEN_COOKIE_NAME, value=cookie, httponly=True, secure=is_heroku(), samesite='Lax', path='/',
-                        max_age=365 * 24 * 60 * 60)
-
-        remember_current_user(user)
-        return resp
-
-    @app.route('/auth/verify', methods=['GET'])
-    def verify_email():
-        username = request.args.get('username', None)
-        token = request.args.get('token', None)
-        if not token:
-            return gettext('token_invalid'), 400
-        if not username:
-            return gettext('username_invalid'), 400
-
-        # Verify that user actually exists
-        user = DATABASE.user_by_username(username)
-        if not user:
-            return gettext('username_invalid'), 403
-
-        # If user is already verified -> re-direct to landing-page anyway
-        if not 'verification_pending' in user:
-            return redirect('/landing-page')
-
-        # Verify the token
-        if token != user['verification_pending']:
-            return gettext('token_invalid'), 403
-
-        # Remove the token from the user
-        DATABASE.update_user(username, {'verification_pending': None})
-
-        # We automatically login the user
-        cookie = make_salt()
-        DATABASE.store_token({'id': cookie, 'username': user['username'], 'ttl': times() + session_length})
-        remember_current_user(user)
-
-        return redirect('/landing-page')
-
-    @app.route('/auth/logout', methods=['POST'])
-    def logout():
-        forget_current_user()
-        if request.cookies.get(TOKEN_COOKIE_NAME):
-            DATABASE.forget_token(request.cookies.get(TOKEN_COOKIE_NAME))
-        return '', 200
-
-    @app.route('/auth/destroy', methods=['POST'])
-    @requires_login
-    def destroy(user):
-        forget_current_user()
-        DATABASE.forget_token(request.cookies.get(TOKEN_COOKIE_NAME))
-        DATABASE.forget_user(user['username'])
-        return '', 200
-
-    @app.route('/auth/destroy_public', methods=['POST'])
-    @requires_login
-    def destroy_public(user):
-        DATABASE.forget_public_profile(user['username'])
-        session.pop('profile_image', None)  # Delete profile image id if existing
-        return '', 200
-
-    @app.route('/auth/change_student_password', methods=['POST'])
-    @requires_login
-    def change_student_password(user):
-        body = request.json
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-        if not isinstance(body.get('password'), str):
-            return gettext('password_invalid'), 400
-        if len(body['password']) < 6:
-            return gettext('password_six'), 400
-
-        if not is_teacher(user):
-            return gettext("password_change_not_allowed"), 400
-        students = DATABASE.get_teacher_students(user['username'])
-        if body['username'] not in students:
-            return gettext("password_change_not_allowed"), 400
-
-        hashed = hash(body['password'], make_salt())
-        DATABASE.update_user(body['username'], {'password': hashed})
-
-        return {'success': gettext("password_change_success")}, 200
-
-    @app.route('/auth/change_password', methods=['POST'])
-    @requires_login
-    def change_password(user):
-        body = request.json
-
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('old_password'), str) or not isinstance(body.get('password'), str):
-            return gettext('password_invalid'), 400
-        if not isinstance(body.get( 'password_repeat'), str):
-            return gettext('repeat_match_password'), 400
-        if len(body['password']) < 6:
-            return gettext('password_six'), 400
-        if body['password'] != body['password_repeat']:
-            return gettext('repeat_match_password'), 400
-
-        # The user object we got from 'requires_login' doesn't have the password, so look that up in the database
-        user = DATABASE.user_by_username(user['username'])
-
-        if not check_password(body['old_password'], user['password']):
-            return gettext('password_invalid'), 403
-
-        hashed = hash(body['password'], make_salt())
-
-        DATABASE.update_user(user['username'], {'password': hashed})
-        # We are not updating the user in the Flask session, because we should not rely on the password in anyway.
-        if not is_testing_request(request):
-            try:
-                send_email_template(template='change_password', email=user['email'], username=user['username'])
-            except:
-                return gettext('mail_error_change_processed'), 400
-
-        return jsonify({'message': gettext('password_updated')}), 200
-
-    @app.route('/profile', methods=['POST'])
-    @requires_login
-    def update_profile(user):
-        body = request.json
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('language'), str) or body.get('language') not in ALL_LANGUAGES.keys():
-            return gettext('language_invalid'), 400
-        if not isinstance(body.get('keyword_language'), str) or body.get('keyword_language') not in ['en', body.get(
-                'language')] or body.get('keyword_language') not in ALL_KEYWORD_LANGUAGES.keys():
-            return gettext('keyword_language_invalid'), 400
-
-        # Mail is a unique field, only mandatory if the user doesn't have a related teacher (and no mail address)
-        user = DATABASE.user_by_username(user['username'])
-        if not user.get('teacher') or 'email' in body:
-            if not isinstance(body.get('email'), str) or not valid_email(body['email']):
-                return gettext('email_invalid'), 400
-
-        # Validations, optional fields
-        if 'birth_year' in body:
-            year = datetime.datetime.now().year
-            try:
-                body['birth_year'] = int(body.get('birth_year'))
-            except ValueError:
-                return gettext('year_invalid').format(**{'current_year': str(year)}), 400
-            if not isinstance(body.get('birth_year'), int) or body['birth_year'] <= 1900 or body['birth_year'] > year:
-                return gettext('year_invalid').format(**{'current_year': str(year)}), 400
-        if 'gender' in body:
-            if body['gender'] not in ["m", "f", "o"]:
-                return gettext('gender_invalid'), 400
-        if 'country' in body:
-            if not body['country'] in COUNTRIES:
-                return gettext('country_invalid'), 400
-
-        resp = {}
-        if 'email' in body:
-            email = body['email'].strip().lower()
-            if email != user.get('email'):
-                exists = DATABASE.user_by_email(email)
-                if exists:
-                    return gettext('exists_email'), 403
-                token = make_salt()
-                hashed_token = hash(token, make_salt())
-                DATABASE.update_user(user['username'], {'email': email, 'verification_pending': hashed_token})
-                # If this is an e2e test, we return the email verification token directly instead of emailing it.
-                if is_testing_request(request):
-                    resp = {'username': user['username'], 'token': hashed_token}
-                else:
-                    try:
-                        send_email_template(template='welcome_verify', email=email,
-                                        link=create_verify_link(user['username'], hashed_token),
-                                        username=user['username'])
-                    except:
-                        # Todo TB: Now we only log to the back-end, would be nice to also return the user some info
-                        # We have two options: return an error at this point (don't process changes)
-                        # Add a notification to the response, still process the changes
-                        print(f"Profile changes processed for {user['username']}, mail sending invalid")
-
-                # We check whether the user is in the Mailchimp list.
-                if not is_testing_request(request) and MAILCHIMP_API_URL:
-                    # We hash the email with md5 to avoid emails with unescaped characters triggering errors
-                    request_path = MAILCHIMP_API_URL + '/members/' + hashlib.md5(
-                        user['email'].encode('utf-8')).hexdigest()
-                    r = requests.get(request_path, headers=MAILCHIMP_API_HEADERS)
-                    # If user is subscribed, we remove the old email from the list and add the new one
-                    if r.status_code == 200:
-                        r = requests.delete(request_path, headers=MAILCHIMP_API_HEADERS)
-                        mailchimp_subscribe_user(email)
-
-        username = user['username']
-
-        updates = {}
-        for field in ['country', 'birth_year', 'gender', 'language', 'keyword_language']:
-            if field in body:
-                updates[field] = body[field]
-            else:
-                updates[field] = None
-        if body.get('agree_third_party'):
-            updates['third_party'] = True
-        else:
-            updates['third_party'] = None
-
-        if updates:
-            DATABASE.update_user(username, updates)
-
-        # Always make sure that the country stored on the public profile is identical to the profile one
-        DATABASE.update_country_public_profile(username, body.get('country', None))
-
-        # We want to check if the user choose a new language, if so -> reload
-        # We can use g.lang for this to reduce the db calls
-        resp['reload'] = False
-        if session['lang'] != body['language'] or session['keyword_lang'] != body['keyword_language']:
-            resp['message'] = gettext('profile_updated_reload')
-            resp['reload'] = True
-        else:
-            resp['message'] = gettext('profile_updated')
-
-        remember_current_user(DATABASE.user_by_username(user['username']))
-        return jsonify(resp)
-
-    @app.route('/profile', methods=['GET'])
-    @requires_login
-    def get_profile(user):
-        # The user object we got from 'requires_login' is not fully hydrated yet. Look up the database user.
-        user = DATABASE.user_by_username(user['username'])
-
-        output = {
-            'username': user['username'],
-            'email': user['email'],
-            'language': user.get('language', 'en')
-        }
-        for field in ['birth_year', 'country', 'gender', 'prog_experience', 'experience_languages', 'third_party']:
-            if field in user:
-                output[field] = user[field]
-        if 'verification_pending' in user:
-            output['verification_pending'] = True
-
-        output['student_classes'] = DATABASE.get_student_classes(user['username'])
-
-        output['session_expires_at'] = timems() + session_length * 1000
-
-        return jsonify(output), 200
-
-    @app.route('/auth/recover', methods=['POST'])
-    def recover():
-        body = request.json
-        # Validations
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-
-        # If username has an @-sign, then it's an email
-        if '@' in body['username']:
-            user = DATABASE.user_by_email(body['username'].strip().lower())
-        else:
-            user = DATABASE.user_by_username(body['username'].strip().lower())
-
-        if not user:
-            return gettext('username_invalid'), 403
-
-        # In this case -> account has a related teacher (and is a student)
-        # We still store the token, but sent the mail to the teacher instead
-        if user.get('teacher') and not user.get('email'):
-            email = DATABASE.user_by_username(user.get('teacher')).get('email')
-        else:
-            email = user['email']
-
-        # Create a token -> use the reset_length value as we don't want the token to live as long as a login one
-        token = make_salt()
-        # Todo TB -> Don't we want to use a hashed token here as well?
-        DATABASE.store_token({'id': token, 'username': user['username'], 'ttl': times() + reset_length})
-
-        if is_testing_request(request):
-            # If this is an e2e test, we return the email verification token directly instead of emailing it.
-            return jsonify({'username': user['username'], 'token': token}), 200
-        else:
-            try:
-                send_email_template(template='recover_password', email=email,
-                                link=create_recover_link(user['username'], token), username=user['username'])
-            except:
-                return gettext('mail_error_change_processed'), 400
-
-            return jsonify({'message': gettext('sent_password_recovery')}), 200
-
-    @app.route('/auth/reset', methods=['POST'])
-    def reset():
-        body = request.json
-        # Validations
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-        if not isinstance(body.get('token'), str):
-            return gettext('token_invalid'), 400
-        if not isinstance(body.get('password'), str):
-            return gettext('password_invalid'), 400
-        if len(body['password']) < 6:
-            return gettext('password_six'), 400
-        if not isinstance(body.get('password_repeat'), str) or body['password'] != body['password_repeat']:
-            return gettext('repeat_match_password'), 400
-
-        token = DATABASE.get_token(body['token'])
-        if not token or body['token'] != token.get('id') or body['username'] != token.get('username'):
-            return gettext('token_invalid'), 403
-
-        hashed = hash(body['password'], make_salt())
-        DATABASE.update_user(body['username'], {'password': hashed})
-        user = DATABASE.user_by_username(body['username'])
-
-        # Delete all tokens of the user -> automatically logout all long-lived sessions
-        DATABASE.delete_all_tokens(body['username'])
-
-        # In this case -> account has a related teacher (and is a student)
-        # We mail the teacher instead
-        if user.get('teacher') and not user.get('email'):
-            email = DATABASE.user_by_username(user.get('teacher')).get('email')
-        else:
-            email = user['email']
-
-        if not is_testing_request(request):
-            try:
-                send_email_template(template='reset_password', email=email, username=user['username'])
-            except:
-                return gettext('mail_error_change_processed'), 400
-
-        return jsonify({'message': gettext('password_resetted')}), 200
-
-    @app.route('/auth/request_teacher', methods=['GET'])
-    @requires_login
-    def request_teacher_account(user):
-        account = DATABASE.user_by_username(user['username'])
-        if account.get('is_teacher'):
-            return gettext('already_teacher'), 400
-        if account.get('teacher_request'):
-            return gettext('already_teacher_request'), 400
-
-        DATABASE.update_user(user['username'], {'teacher_request': True})
-        return jsonify({'message': gettext('teacher_account_success')}), 200
-
-    # *** ADMIN ROUTES ***
-
-    @app.route('/admin/markAsTeacher', methods=['POST'])
-    def mark_as_teacher():
-        user = current_user()
-        if not is_admin(user) and not is_testing_request(request):
-            return utils.error_page(error=403, ui_message=gettext('unauthorized'))
-
-        body = request.json
-
-        # Validations
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-        if not isinstance(body.get('is_teacher'), bool):
-            return gettext('teacher_invalid'), 400
-
-        user = DATABASE.user_by_username(body['username'].strip().lower())
-
-        if not user:
-            return gettext('username_invalid'), 400
-
-        is_teacher_value = 1 if body['is_teacher'] else 0
-        update_is_teacher(user, is_teacher_value)
-
-        # Todo TB feb 2022 -> Return the success message here instead of fixing in the front-end
-        return '', 200
-
-    @app.route('/admin/changeUserEmail', methods=['POST'])
-    @requires_admin
-    def change_user_email(user):
-        body = request.json
-
-        # Validations
-        if not isinstance(body, dict):
-            return gettext('ajax_error'), 400
-        if not isinstance(body.get('username'), str):
-            return gettext('username_invalid'), 400
-        if not isinstance(body.get('email'), str) or not valid_email(body['email']):
-            return gettext('email_invalid'), 400
-
-        user = DATABASE.user_by_username(body['username'].strip().lower())
-
-        if not user:
-            return gettext('email_invalid'), 400
-
-        token = make_salt()
-        hashed_token = hash(token, make_salt())
-
-        # We assume that this email is not in use by any other users. In other words, we trust the admin to enter a valid, not yet used email address.
-        DATABASE.update_user(user['username'], {'email': body['email'], 'verification_pending': hashed_token})
-
-        # If this is an e2e test, we return the email verification token directly instead of emailing it.
-        if is_testing_request(request):
-            resp = {'username': user['username'], 'token': hashed_token}
-        else:
-            try:
-                send_email_template(template='welcome_verify', email=body['email'],
-                                    link=create_verify_link(user['username'], hashed_token),
-                                    username=user['username'])
-            except:
-                return gettext('mail_error_change_processed'), 400
-
-        return {}, 200
-
-    @app.route('/admin/getUserTags', methods=['POST'])
-    @requires_admin
-    def get_user_tags(user):
-        body = request.json
-        user = DATABASE.get_public_profile_settings(body['username'].strip().lower())
-        if not user:
-            return "User doesn't have a public profile", 400
-        return {'tags': user.get('tags', [])}, 200
-
-    @app.route('/admin/updateUserTags', methods=['POST'])
-    @requires_admin
-    def update_user_tags(user):
-        body = request.json
-        user = DATABASE.get_public_profile_settings(body['username'].strip().lower())
-        if not user:
-            return "User doesn't have a public profile", 400
-
-        tags = []
-        if "admin" in user.get('tags', []):
-            tags.append("admin")
-        if "teacher" in user.get('tags', []):
-            tags.append("teacher")
-        if body.get("certified"):
-            tags.append("certified_teacher")
-        if body.get("distinguished"):
-            tags.append("distinguished_user")
-        if body.get("contributor"):
-            tags.append("contributor")
-
-        user['tags'] = tags
-        DATABASE.update_public_profile(user['username'], user)
-        return {}, 200
-
 # Turn off verbose logs from boto/SES, thanks to https://github.com/boto/boto3/issues/521
 import logging
 
@@ -879,7 +218,6 @@ for name in logging.Logger.manager.loggerDict.keys():
 
 # https://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-using-sdk-python.html
 email_client = boto3.client('ses', region_name=config['email']['region'])
-
 
 @querylog.timed
 def send_email(recipient, subject, body_plain, body_html):
@@ -952,6 +290,36 @@ def send_email_template(template, email, link=None, username=gettext('user')):
 
     send_email(email, subject, body_plain, body_html)
 
+def store_new_student_account(db, account, teacher_username):
+    username, hashed, hashed_token = prepare_user_db(account['username'], account['password'])
+    user = {
+        'username': username,
+        'password': hashed,
+        'language': account['language'],
+        'keyword_language': account['keyword_language'],
+        'created': timems(),
+        'teacher': teacher_username,
+        'verification_pending': hashed_token,
+        'last_login': timems()
+    }
+    db.store_user(user)
+    return user
+
+def prepare_user_db(username, password):
+    hashed = hash(password, make_salt())
+
+    token = make_salt()
+    hashed_token = hash(token, make_salt())
+    username = username.strip().lower()
+
+    return username, hashed, hashed_token
+
+
+def create_verify_link(username, token):
+    email = email_base_url() + '/auth/verify?username='
+    email += urllib.parse.quote_plus(username) + '&token=' + urllib.parse.quote_plus(token)
+    return email
+
 
 def email_base_url():
     """Return the base URL for the current site, without trailing slash.
@@ -967,3 +335,8 @@ def email_base_url():
     if from_env:
         return from_env.rstrip('/')
     return request.host_url
+
+def create_recover_link(username, token):
+    email = email_base_url() + '/reset?username='
+    email += urllib.parse.quote_plus(username) + '&token=' + urllib.parse.quote_plus(token)
+    return email
