+++ website/parsons.py
@@ -1,23 +1,17 @@
-import copy
-import random
-
-from flask import g, jsonify
+from flask import jsonify, g
 from flask_babel import gettext
+from .website_module import WebsiteModule, route
 
+class ParsonsModule(WebsiteModule):
+    def __init__(self, parsons):
+        super().__init__('parsons', __name__, url_prefix='/parsons')
 
-def routes(app, database, achievements, parsons):
-    global DATABASE
-    global ACHIEVEMENTS
-    global PARSONS
+        self.parsons = parsons
 
-    DATABASE = database
-    ACHIEVEMENTS = achievements
-    PARSONS = parsons
-
-    @app.route('/parsons/get-exercise/<int:level>/<int:exercise>', methods=['GET'])
-    def get_parsons_exercise(level, exercise):
-        if exercise > PARSONS[g.lang].get_highest_exercise_level(level) or exercise < 1:
+    @route('/get-exercise/<int:level>/<int:exercise>', methods=['GET'])
+    def get_parsons_exercise(self, level, exercise):
+        if exercise > self.parsons[g.lang].get_highest_exercise_level(level) or exercise < 1:
             return gettext('exercise_doesnt_exist'), 400
 
-        exercise = PARSONS[g.lang].get_parsons_data_for_level_exercise(level, exercise, g.keyword_lang)
+        exercise = self.parsons[g.lang].get_parsons_data_for_level_exercise(level, exercise, g.keyword_lang)
         return jsonify(exercise), 200
