+++ website/programs.py
@@ -1,50 +1,53 @@
+from flask import g, request, jsonify
 from flask_babel import gettext
 import hedy
 from config import config
+from .achievements import Achievements
+from .database import Database
 from website.auth import requires_login, current_user, is_admin, send_email, email_base_url, requires_admin
 import utils
 import uuid
-from flask import g, request, jsonify
 
+from .website_module import WebsiteModule, route
 
-def routes(app, database, achievements):
-    global DATABASE
-    global ACHIEVEMENTS
-    DATABASE = database
-    ACHIEVEMENTS = achievements
+class ProgramsModule(WebsiteModule):
+    def __init__(self, db: Database, achievements: Achievements):
+        super().__init__('programs', __name__, url_prefix='/programs')
+        self.db = db
+        self.achievements = achievements
 
-    @app.route('/programs_list', methods=['GET'])
+    @route('/list', methods=['GET'])
     @requires_login
-    def list_programs(user):
-        return {'programs': DATABASE.programs_for_user(user['username']).records}
+    def list_programs(self, user):
+        return {'programs': self.db.programs_for_user(user['username']).records}
 
-    @app.route('/programs/delete/', methods=['POST'])
+    @route('/delete/', methods=['POST'])
     @requires_login
-    def delete_program(user):
+    def delete_program(self, user):
         body = request.json
         if not isinstance(body.get('id'), str):
             return 'program id must be a string', 400
 
-        result = DATABASE.program_by_id(body['id'])
+        result = self.db.program_by_id(body['id'])
 
         if not result or (result['username'] != user['username'] and not is_admin(user)):
             return "", 404
-        DATABASE.delete_program_by_id(body['id'])
-        DATABASE.increase_user_program_count(user['username'], -1)
+        self.db.delete_program_by_id(body['id'])
+        self.db.increase_user_program_count(user['username'], -1)
 
         # This only happens in the situation were a user deletes their favourite program -> Delete from public profile
-        public_profile = DATABASE.get_public_profile_settings(current_user()['username'])
+        public_profile = self.db.get_public_profile_settings(current_user()['username'])
         if public_profile and 'favourite_program' in public_profile and public_profile['favourite_program'] == body['id']:
-            DATABASE.set_favourite_program(user['username'], None)
+            self.db.set_favourite_program(user['username'], None)
 
-        achievement = ACHIEVEMENTS.add_single_achievement(user['username'], "do_you_have_copy")
+        achievement = self.achievements.add_single_achievement(user['username'], "do_you_have_copy")
         resp = {'message': gettext('delete_success')}
         if achievement:
             resp['achievement'] = achievement
         return jsonify(resp)
 
-    @app.route('/programs/duplicate-check', methods=['POST'])
-    def check_duplicate_program():
+    @route('/duplicate-check', methods=['POST'])
+    def check_duplicate_program(self):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
@@ -54,15 +57,15 @@ def routes(app, database, achievements):
         if not current_user()['username']:
             return gettext('save_prompt'), 403
 
-        programs = DATABASE.programs_for_user(current_user()['username'])
+        programs = self.db.programs_for_user(current_user()['username'])
         for program in programs:
             if program['name'] == body['name']:
                 return jsonify({'duplicate': True, 'message': gettext('overwrite_warning')})
         return jsonify({'duplicate': False})
 
-    @app.route('/programs', methods=['POST'])
+    @route('/programs', methods=['POST'])
     @requires_login
-    def save_program(user):
+    def save_program(self, user):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
@@ -89,7 +92,7 @@ def routes(app, database, achievements):
         # We check if a program with a name `xyz` exists in the database for the username.
         # It'd be ideal to search by username & program name, but since DynamoDB doesn't allow searching for two indexes at the same time, this would require to create a special index to that effect, which is cumbersome.
         # For now, we bring all existing programs for the user and then search within them for repeated names.
-        programs = DATABASE.programs_for_user(user['username']).records
+        programs = self.db.programs_for_user(user['username']).records
         program_id = uuid.uuid4().hex
         program_public = body.get('shared')
         overwrite = False
@@ -119,21 +122,21 @@ def routes(app, database, achievements):
         if 'adventure_name' in body:
             stored_program['adventure_name'] = body['adventure_name']
 
-        DATABASE.store_program(stored_program)
+        self.db.store_program(stored_program)
         if not overwrite:
-            DATABASE.increase_user_program_count(user['username'])
-        DATABASE.increase_user_save_count(user['username'])
-        ACHIEVEMENTS.increase_count("saved")
+            self.db.increase_user_program_count(user['username'])
+        self.db.increase_user_save_count(user['username'])
+        self.achievements.increase_count("saved")
 
-        if ACHIEVEMENTS.verify_save_achievements(user['username'],
-                                                 'adventure_name' in body and len(body['adventure_name']) > 2):
+        if self.achievements.verify_save_achievements(user['username'],
+                                                    'adventure_name' in body and len(body['adventure_name']) > 2):
             return jsonify(
-                {'message': gettext('save_success_detail'), 'name': body['name'], 'id': program_id, "achievements": ACHIEVEMENTS.get_earned_achievements()})
+                {'message': gettext('save_success_detail'), 'name': body['name'], 'id': program_id, "achievements": self.achievements.get_earned_achievements()})
         return jsonify({'message': gettext('save_success_detail'), 'name': body['name'], 'id': program_id})
 
-    @app.route('/programs/share', methods=['POST'])
+    @route('/share', methods=['POST'])
     @requires_login
-    def share_unshare_program(user):
+    def share_unshare_program(self, user):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
@@ -142,18 +145,18 @@ def routes(app, database, achievements):
         if not isinstance(body.get('public'), bool):
             return 'public must be a boolean', 400
 
-        result = DATABASE.program_by_id(body['id'])
+        result = self.db.program_by_id(body['id'])
         if not result or result['username'] != user['username']:
             return 'No such program!', 404
 
         # This only happens in the situation were a user un-shares their favourite program -> Delete from public profile
-        public_profile = DATABASE.get_public_profile_settings(current_user()['username'])
+        public_profile = self.db.get_public_profile_settings(current_user()['username'])
         if public_profile and 'favourite_program' in public_profile and public_profile['favourite_program'] == body[
             'id']:
-            DATABASE.set_favourite_program(user['username'], None)
+            self.db.set_favourite_program(user['username'], None)
 
-        DATABASE.set_program_public_by_id(body['id'], bool(body['public']))
-        achievement = ACHIEVEMENTS.add_single_achievement(user['username'], "sharing_is_caring")
+        self.db.set_program_public_by_id(body['id'], bool(body['public']))
+        achievement = self.achievements.add_single_achievement(user['username'], "sharing_is_caring")
 
         resp = {'id': body['id']}
         if bool(body['public']):
@@ -164,48 +167,48 @@ def routes(app, database, achievements):
             resp['achievement'] = achievement
         return jsonify(resp)
 
-    @app.route('/programs/submit', methods=['POST'])
+    @route('/submit', methods=['POST'])
     @requires_login
-    def submit_program(user):
+    def submit_program(self, user):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
         if not isinstance(body.get('id'), str):
             return 'id must be a string', 400
 
-        result = DATABASE.program_by_id(body['id'])
+        result = self.db.program_by_id(body['id'])
         if not result or result['username'] != user['username']:
             return 'No such program!', 404
 
-        DATABASE.submit_program_by_id(body['id'])
-        DATABASE.increase_user_submit_count(user['username'])
-        ACHIEVEMENTS.increase_count("submitted")
+        self.db.submit_program_by_id(body['id'])
+        self.db.increase_user_submit_count(user['username'])
+        self.achievements.increase_count("submitted")
 
-        if ACHIEVEMENTS.verify_submit_achievements(user['username']):
-            return jsonify({"achievements": ACHIEVEMENTS.get_earned_achievements()})
+        if self.achievements.verify_submit_achievements(user['username']):
+            return jsonify({"achievements": self.achievements.get_earned_achievements()})
         return jsonify({})
 
-    @app.route('/programs/set_favourite', methods=['POST'])
+    @route('/set_favourite', methods=['POST'])
     @requires_login
-    def set_favourite_program(user):
+    def set_favourite_program(self, user):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
         if not isinstance(body.get('id'), str):
             return 'id must be a string', 400
 
-        result = DATABASE.program_by_id(body['id'])
+        result = self.db.program_by_id(body['id'])
         if not result or result['username'] != user['username']:
             return 'No such program!', 404
 
-        if DATABASE.set_favourite_program(user['username'], body['id']):
+        if self.db.set_favourite_program(user['username'], body['id']):
             return jsonify({'message': gettext('favourite_success')})
         else:
             return "You can't set a favourite program without a public profile", 400
 
-    @app.route('/programs/set_hedy_choice', methods=['POST'])
+    @route('/set_hedy_choice', methods=['POST'])
     @requires_admin
-    def set_hedy_choice(user):
+    def set_hedy_choice(self, user):
         body = request.json
         if not isinstance(body, dict):
             return 'body must be an object', 400
@@ -216,28 +219,28 @@ def routes(app, database, achievements):
 
         favourite = True if body.get('favourite') == 1 else False
 
-        result = DATABASE.program_by_id(body['id'])
+        result = self.db.program_by_id(body['id'])
         if not result:
             return 'No such program!', 404
 
-        DATABASE.set_program_as_hedy_choice(body['id'], favourite)
+        self.db.set_program_as_hedy_choice(body['id'], favourite)
         if favourite:
             return jsonify({'message': 'Program successfully set as a "Hedy choice" program.'}), 200
         return jsonify({'message': 'Program successfully removed as a "Hedy choice" program.'}), 200
 
-    @app.route('/programs/report', methods=['POST'])
+    @route('/report', methods=['POST'])
     @requires_login
-    def report_program(user):
+    def report_program(self, user):
         body = request.json
 
         # Make sure the program actually exists and is public
-        program = DATABASE.program_by_id(body.get('id'))
+        program = self.db.program_by_id(body.get('id'))
         if not program or program.get('public') != 1:
             return gettext('report_failure'), 400
 
         link = email_base_url() + '/hedy/' + body.get('id') + '/view'
         send_email(config['email']['sender'], "The following program is reported by " + user['username'], link,
-                   '<a href="' + link + '">Program link</a>')
+                    '<a href="' + link + '">Program link</a>')
 
         return {'message': gettext('report_success')}, 200
